+++ domain/src/main/java/com/sephoratest/domain/products/models/Brand.kt
@@ -1,0 +1,6 @@
+package com.sephoratest.domain.products.models
+
+data class Brand(
+    val id: String,
+    val name: String
+)