+++ domain/src/main/java/com/sephoratest/domain/products/models/ProductImages.kt
@@ -1,0 +1,6 @@
+package com.sephoratest.domain.products.models
+
+data class ProductImages(
+    val smallUrl: String,
+    val largeUrl: String,
+)