+++ domain/src/main/java/com/sephoratest/domain/products/repositories/ProductRepository.kt
@@ -1,0 +1,8 @@
+package com.sephoratest.domain.products.repositories
+
+import com.sephoratest.domain.products.models.Product
+
+interface ProductRepository {
+
+    suspend fun getAll(): Result<List<Product>>
+}