+++ domain/src/main/java/com/sephoratest/domain/products/usecases/GetAllProductsUseCase.kt
@@ -1,0 +1,9 @@
+package com.sephoratest.domain.products.usecases
+
+import com.sephoratest.domain.products.repositories.ProductRepository
+
+abstract class GetAllProductsUseCase(private val productRepository: ProductRepository) {
+
+    suspend operator fun invoke() =
+        productRepository.getAll()
+}