+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initiated CHANGELOG.md, README.md
 - Initiated domain and core modules
+- Implemented domain layer for products
 
 ### Changed